Total Purchase Amount issue fixed

+++ additional_addons/kingswood/kw_stock.py
@@ -4148,7 +4148,7 @@ class stock_picking_out(osv.osv):
                 if sub_part_ids:
                     sub_part_ids = sub_part_ids[0]
                     sub_part = sub_part_obj.browse(cr, uid, sub_part_ids)
-                    if sub_part.total_purchase >= float(700000):
+                    if sub_part.total_purchase >= float(1980000):
                         raise osv.except_osv(_('Warning'),_('Total Purcase is exceeded for the selected Sub Facilitator.'))
 
             if 'move_lines' in vals:
